Add Statistic Object

+++ src/MYRAY.Business/DTOs/Statistic/StatisticDetail.cs
@@ -2,7 +2,7 @@ namespace MYRAY.Business.DTOs.Statistic;
 
 public class StatisticDetail
 {
-    public int TotalMoney { get; set; }
+    public double TotalMoney { get; set; }
     public int TotalJobPost { get; set; }
     public int TotalLandowner { get; set; }
     public int TotalFarmer { get; set; }
